fix(encoder): setBitrate 파라미터를 gint 로 넓혀 절단을 막는다
Codex 리뷰 지적(P2). setBitrate 의 파라미터가 guint16 인데 호출부는
charArrayToInt() 가 돌려준 gint 를 넘긴다. 종전에는 `set bps` 검증의
`key > 9999` 가 우연히 이 절단을 막고 있었는데, 앞 커밋에서 상한을 없애면서
65535 초과 값이 그대로 guint16 으로 들어가게 됐다.

  set bps rec 0 70000  ->  70000 & 0xFFFF = 4464

"0 이상이면 통과" 라는 의도와 정반대로, 큰 값이 더 낮은 비트레이트로
프로그래밍된다. 세 Bin 의 선언·정의를 gint 로 넓혀 값이 그대로 인코더까지
도달하게 하고, 상한은 wrapper 의 VPU_ENC_MAX_BITRATE(60000 kbps) 클램프에
맡긴다.

setGop 은 guint16 을 유지한다 — MAX_GOP_SIZE 가 300 이라 절단이 불가능하다.

+++ encoderBin.cpp
@@ -45,7 +45,7 @@ void EncoderBin::getBitrate()
      g_print("rec ch%d get bitrate : %d\n", encData.ch, bps);
 }
 
-void EncoderBin::setBitrate(guint16 data)
+void EncoderBin::setBitrate(gint data)
 {
     gint bps;
